requirements.txt
pyTelegramBotAPI
mysql-connector-python
